refactor MooncakeStoreConfig and connector init for clarity
