+++ lmcache/v1/storage_backend/connector/mooncakestore_connector.py
@@ -42,27 +42,29 @@ class MooncakeStoreConfig:
     dummy_server_address: str = ""
 
     @staticmethod
+    def _from_dict(d: dict) -> "MooncakeStoreConfig":
+        """Build a MooncakeStoreConfig from a plain dict."""
+        return MooncakeStoreConfig(
+            local_hostname=d.get("local_hostname", ""),
+            metadata_server=d.get("metadata_server", ""),
+            global_segment_size=d.get("global_segment_size", 3355443200),
+            local_buffer_size=d.get("local_buffer_size", 1073741824),
+            protocol=d.get("protocol", "tcp"),
+            device_name=d.get("device_name", ""),
+            master_server_address=d.get("master_server_address", ""),
+            transfer_timeout=d.get("transfer_timeout", 1),
+            storage_root_dir=d.get("storage_root_dir", ""),
+            prefer_local_alloc=d.get("mooncake_prefer_local_alloc", False),
+            use_dummy_client=d.get("use_dummy_client", False),
+            dummy_server_address=d.get("dummy_server_address", ""),
+        )
+
+    @staticmethod
     def from_file(file_path: str) -> "MooncakeStoreConfig":
         """Load the config from a JSON file."""
         with open(file_path) as fin:
             config = json.load(fin)
-        # Read Mooncake-specific knob
-        prefer_local_alloc = config.get("mooncake_prefer_local_alloc", False)
-
-        return MooncakeStoreConfig(
-            local_hostname=config.get("local_hostname", ""),
-            metadata_server=config.get("metadata_server", ""),
-            global_segment_size=config.get("global_segment_size", 3355443200),
-            local_buffer_size=config.get("local_buffer_size", 1073741824),
-            protocol=config.get("protocol", "tcp"),
-            device_name=config.get("device_name", ""),
-            master_server_address=config.get("master_server_address", ""),
-            transfer_timeout=config.get("transfer_timeout", 1),
-            storage_root_dir=config.get("storage_root_dir", ""),
-            prefer_local_alloc=prefer_local_alloc,
-            use_dummy_client=config.get("use_dummy_client", False),
-            dummy_server_address=config.get("dummy_server_address", ""),
-        )
+        return MooncakeStoreConfig._from_dict(config)
 
     @staticmethod
     def load_from_env() -> "MooncakeStoreConfig":
@@ -78,27 +80,11 @@ class MooncakeStoreConfig:
     def load_from_lmcache_config(
         config: "LMCacheEngineConfig",
     ) -> "MooncakeStoreConfig":
-        """Load config from a file specified in the environment variable."""
+        """Load config from the lmcache engine config's extra_config dict."""
         extra_config = config.extra_config
         if extra_config is None:
             raise ValueError("The extra config is not set.")
-        # Read Mooncake-specific knob
-        prefer_local_alloc = extra_config.get("mooncake_prefer_local_alloc", False)
-
-        return MooncakeStoreConfig(
-            local_hostname=extra_config.get("local_hostname", ""),
-            metadata_server=extra_config.get("metadata_server", ""),
-            global_segment_size=extra_config.get("global_segment_size", 3355443200),
-            local_buffer_size=extra_config.get("local_buffer_size", 1073741824),
-            protocol=extra_config.get("protocol", "tcp"),
-            device_name=extra_config.get("device_name", ""),
-            master_server_address=extra_config.get("master_server_address", ""),
-            transfer_timeout=extra_config.get("transfer_timeout", 1),
-            storage_root_dir=extra_config.get("storage_root_dir", ""),
-            prefer_local_alloc=prefer_local_alloc,
-            use_dummy_client=extra_config.get("use_dummy_client", False),
-            dummy_server_address=extra_config.get("dummy_server_address", ""),
-        )
+        return MooncakeStoreConfig._from_dict(extra_config)
 
 
 class MooncakestoreConnector(RemoteConnector):
@@ -171,105 +157,9 @@ class MooncakestoreConnector(RemoteConnector):
                 )
 
             if self.config.use_dummy_client:
-                # Dummy client mode: use setup_dummy with shared memory + IPC
-                if not hasattr(self.store, "setup_dummy"):
-                    raise RuntimeError(
-                        "The installed mooncake version does not support "
-                        "dummy client mode (setup_dummy). Please upgrade "
-                        "mooncake or set use_dummy_client to False."
-                    )
-                if not self.config.dummy_server_address:
-                    raise ValueError(
-                        "dummy_server_address must be provided when "
-                        "use_dummy_client is True."
-                    )
-                # Skip NUMA binding: RDMA runs in separate Real Client process
-                logger.info("Skipping NUMA binding in dummy client mode")
-                logger.info(
-                    "Using dummy client mode with server address: %s",
-                    self.config.dummy_server_address,
-                )
-                self.store.setup_dummy(
-                    self.config.global_segment_size,
-                    self.config.local_buffer_size,
-                    self.config.dummy_server_address,
-                )
-                logger.info("Mooncake dummy client setup completed successfully")
-
-                # Replace pin_allocator with shm-backed memory
-                allocator = local_cpu_backend.memory_allocator
-                if hasattr(allocator, "pin_allocator") and hasattr(
-                    allocator.pin_allocator, "buffer"
-                ):
-                    buffer_size = allocator.pin_allocator.buffer.numel()
-
-                    # Allocate shm from mooncake's memory pool
-                    shm_ptr = self.store.alloc_from_mem_pool(buffer_size)
-                    if not shm_ptr:
-                        raise RuntimeError(
-                            "Failed to allocate shared memory from mooncake mem pool"
-                        )
-
-                    # Delegate all pin/wrap/rebuild to the allocator
-                    allocator.replace_buffer(shm_ptr, buffer_size)  # type: ignore[attr-defined]
-
-                    logger.info(
-                        "Replaced pin_allocator with shm buffer: ptr=%s, size=%d",
-                        hex(shm_ptr),
-                        buffer_size,
-                    )
+                self._setup_dummy_client(local_cpu_backend)
             else:
-                # Real client mode: NUMA binding + standard setup
-                try:
-                    numa_mapping = getattr(
-                        local_cpu_backend.memory_allocator, "numa_mapping", None
-                    )
-                    if numa_mapping is None and lmcache_config is not None:
-                        numa_mapping = NUMADetector.get_numa_mapping(lmcache_config)
-
-                    if numa_mapping:
-                        current_device_id = torch.cuda.current_device()
-                        gpu_to_numa = getattr(numa_mapping, "gpu_to_numa_mapping", {})
-                        numa_id = gpu_to_numa.get(current_device_id)
-                        logger.info(
-                            f"NUMA mapping detected (pre-Mooncake setup): {gpu_to_numa}"
-                        )
-                        try:
-                            # Third Party
-                            from mooncake.store import bind_to_numa_node
-
-                            if numa_id is not None:
-                                bind_to_numa_node(numa_id)
-                                logger.info(
-                                    f"GPU {current_device_id}, "
-                                    f"NUMA node {numa_id} binding done"
-                                )
-                            else:
-                                logger.info(
-                                    "NUMA mapping not found "
-                                    f"for GPU {current_device_id}"
-                                )
-                        except ImportError:
-                            logger.warning(
-                                "unable to import bind_to_numa_node from mooncake.store"
-                            )
-                    else:
-                        logger.info("NUMA mapping unavailable or disabled")
-                except Exception as e:
-                    logger.warning(
-                        f"Failed to determine NUMA mapping before Mooncake setup: {e}"
-                    )
-
-                self.store.setup(
-                    self.config.local_hostname,
-                    self.config.metadata_server,
-                    self.config.global_segment_size,
-                    self.config.local_buffer_size,
-                    self.config.protocol,
-                    self.config.device_name,
-                    self.config.master_server_address,
-                )
-                logger.info("Mooncake store setup completed successfully")
+                self._setup_real_client(local_cpu_backend, lmcache_config)
 
         except ValueError as e:
             logger.error("Configuration loading failed: %s", e)
@@ -294,6 +184,106 @@ class MooncakestoreConnector(RemoteConnector):
         self._register_cpu_buffer()
 
         logger.info("MooncakeConnector initialized successfully.")
+
+    def _setup_dummy_client(self, local_cpu_backend: LocalCPUBackend):
+        """Set up dummy client mode: shared memory + IPC with Real Client."""
+        if not hasattr(self.store, "setup_dummy"):
+            raise RuntimeError(
+                "The installed mooncake version does not support "
+                "dummy client mode (setup_dummy). Please upgrade "
+                "mooncake or set use_dummy_client to False."
+            )
+        if not self.config.dummy_server_address:
+            raise ValueError(
+                "dummy_server_address must be provided when use_dummy_client is True."
+            )
+        logger.info("Skipping NUMA binding in dummy client mode")
+        logger.info(
+            "Using dummy client mode with server address: %s",
+            self.config.dummy_server_address,
+        )
+        self.store.setup_dummy(
+            self.config.global_segment_size,
+            self.config.local_buffer_size,
+            self.config.dummy_server_address,
+        )
+        logger.info("Mooncake dummy client setup completed successfully")
+
+        # Replace pin_allocator with shm-backed memory
+        allocator = local_cpu_backend.memory_allocator
+        if hasattr(allocator, "pin_allocator") and hasattr(
+            allocator.pin_allocator, "buffer"
+        ):
+            buffer_size = allocator.pin_allocator.buffer.numel()
+
+            shm_ptr = self.store.alloc_from_mem_pool(buffer_size)
+            if not shm_ptr:
+                raise RuntimeError(
+                    "Failed to allocate shared memory from mooncake mem pool"
+                )
+
+            allocator.replace_buffer(shm_ptr, buffer_size)  # type: ignore[attr-defined]
+
+            logger.info(
+                "Replaced pin_allocator with shm buffer: ptr=%s, size=%d",
+                hex(shm_ptr),
+                buffer_size,
+            )
+
+    def _setup_real_client(
+        self,
+        local_cpu_backend: LocalCPUBackend,
+        lmcache_config: Optional[LMCacheEngineConfig],
+    ):
+        """Set up real client mode: NUMA binding + standard Mooncake setup."""
+        try:
+            numa_mapping = getattr(
+                local_cpu_backend.memory_allocator, "numa_mapping", None
+            )
+            if numa_mapping is None and lmcache_config is not None:
+                numa_mapping = NUMADetector.get_numa_mapping(lmcache_config)
+
+            if numa_mapping:
+                current_device_id = torch.cuda.current_device()
+                gpu_to_numa = getattr(numa_mapping, "gpu_to_numa_mapping", {})
+                numa_id = gpu_to_numa.get(current_device_id)
+                logger.info(
+                    f"NUMA mapping detected (pre-Mooncake setup): {gpu_to_numa}"
+                )
+                try:
+                    # Third Party
+                    from mooncake.store import bind_to_numa_node
+
+                    if numa_id is not None:
+                        bind_to_numa_node(numa_id)
+                        logger.info(
+                            f"GPU {current_device_id}, NUMA node {numa_id} binding done"
+                        )
+                    else:
+                        logger.info(
+                            f"NUMA mapping not found for GPU {current_device_id}"
+                        )
+                except ImportError:
+                    logger.warning(
+                        "unable to import bind_to_numa_node from mooncake.store"
+                    )
+            else:
+                logger.info("NUMA mapping unavailable or disabled")
+        except Exception as e:
+            logger.warning(
+                f"Failed to determine NUMA mapping before Mooncake setup: {e}"
+            )
+
+        self.store.setup(
+            self.config.local_hostname,
+            self.config.metadata_server,
+            self.config.global_segment_size,
+            self.config.local_buffer_size,
+            self.config.protocol,
+            self.config.device_name,
+            self.config.master_server_address,
+        )
+        logger.info("Mooncake store setup completed successfully")
 
     def _register_cpu_buffer(self):
         """Register CPU buffer for zero-copy operations."""
